Backend/requirements.txt
# FastAPI Backend Requirements - Irama1Asia

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Upload & Storage
aiofiles==23.2.1
pillow==10.1.0
boto3==1.34.0

# Real-time Features
websockets==12.0
redis==5.0.1
celery==5.3.4

# Email & SMS
fastapi-mail==1.4.1
twilio==8.10.0

# Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# CORS & Middleware
fastapi-cors==0.0.6

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
bcrypt==4.1.2
python-dateutil==2.8.2